+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #191 ExcelLoader.skipRows(int)
 * #192 Avoid "arg + vararg" methods
 * #195 Bump org.apache.avro:avro from 1.10.1 to 1.11.3
+* #196 StrExp.contains()
 
 ## Release 0.17